% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-ui.R
\name{add_dependency}
\alias{add_dependency}
\title{Add a vdiffr dependency}
\usage{
add_dependency(deps)
}
\arguments{
\item{deps}{A vector containing the names of the packages for which
a dependency should be added.}
}
\description{
It is useful to record the version number of all the packages on
which your visual test cases depend. A note containing a version
number is added to the \code{DESCRIPTION} file for each dependency.
Dependencies on svglite and ggplot2 are automatically added. In
addition, \code{add_dependency()} can be called in any testthat file to
manually add a dependency to a package.
}
\keyword{internal}
