% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{collect_cases}
\alias{collect_cases}
\alias{collect_mismatched_cases}
\alias{collect_new_cases}
\alias{validate_cases}
\title{Collect and validate cases}
\usage{
collect_cases(package = ".", filter = NULL)

collect_new_cases(package = ".")

collect_mismatched_cases(package = ".")

validate_cases(cases = collect_new_cases())
}
\arguments{
\item{package}{Package description, can be path or package
name. See \code{\link[devtools]{as.package}} for more information.}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{cases}{A \code{cases} object returned by one of the
collecting functions.}
}
\description{
These functions are mainly intended for internal use by
\code{\link{manage_cases}()}. They are useful to programmatically
collect and validate cases.
}
\examples{
\dontrun{
new_cases <- collect_new_cases()
validate_cases(new_cases)
}
}
\seealso{
\code{\link{manage_cases}()}
}

