% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/widgets.R
\name{htmlwidgets}
\alias{htmlwidgets}
\alias{widget_diff}
\alias{widget_slide}
\alias{widget_toggle}
\title{HTML Widgets for graphical comparison}
\usage{
widget_toggle(before, after, width = NULL, height = NULL)

widget_slide(before, after, width = NULL, height = NULL)

widget_diff(before, after, width = NULL, height = NULL)
}
\arguments{
\item{before}{Path to the picture that is taken as reference.}

\item{after}{Path to the picture against which the reference is
compared.}

\item{width}{Fixed width for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}

\item{height}{Fixed height for widget (in css units). The default is
\code{NULL}, which results in intelligent automatic sizing based on the
widget's container.}
}
\description{
HTML Widgets for graphical comparison
}

