% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-reporter.R
\name{last_collection_error}
\alias{last_collection_error}
\title{Print last error that occurred during collection}
\usage{
last_collection_error()
}
\description{
Print last error that occurred during collection
}
