% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases-ui.R
\name{manage_cases}
\alias{manage_cases}
\title{Manage visual test cases with a Shiny app}
\usage{
manage_cases(package = ".", filter = NULL)
}
\arguments{
\item{package}{Package description, can be path or package
name. See \code{\link[devtools]{as.package}} for more information.}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed.  Matching will take on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}
}
\description{
Manage visual test cases with a Shiny app
}
\seealso{
\code{\link{vdiffrAddin}()}, \code{\link{collect_cases}()}
  and \code{\link{validate_cases}()}
}

