% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{write_svg}
\alias{write_svg}
\title{Default SVG writer}
\usage{
write_svg(plot, file, title = "")
}
\arguments{
\item{plot}{A plot object to convert to SVG. Can be a ggplot2 object,
a \link[grDevices:recordplot]{recorded plot}, or any object with a
\link[base:print]{print()} method.}

\item{file}{The file to write the SVG to.}

\item{title}{An optional title for the test case.}
}
\description{
This is the default SVG writer for vdiffr test cases. It uses
embedded versions of \href{https://svglite.r-lib.org}{svglite},
\href{https://harfbuzz.github.io/}{harfbuzz}, and the Liberation and
Symbola fonts in order to create deterministic SVGs.
}
