% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{collect_cases}
\alias{collect_cases}
\alias{collect_new_cases}
\alias{collect_mismatched_cases}
\alias{collect_orphaned_cases}
\title{Collect cases}
\usage{
collect_cases(package = ".", filter = NULL, invert = FALSE)

collect_new_cases(package = ".")

collect_mismatched_cases(package = ".")

collect_orphaned_cases(package = ".")
}
\arguments{
\item{package}{Package description, can be path or package
name. See \code{\link[devtools:as.package]{devtools::as.package()}} for more information.}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed. Matching be performed on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{invert}{should the regexp supplied to \code{filter} be inverted? Defaults to \code{FALSE}.}
}
\value{
A \code{cases} object. \code{collect_new_cases()},
\code{collect_mismatched_cases()} and \code{collect_orphaned_cases()}
return a filtered \code{cases} object.
}
\description{
These functions are mainly intended for internal use by
\code{\link[=manage_cases]{manage_cases()}}. They are useful to programmatically collect
cases.
}
\seealso{
\code{\link[=manage_cases]{manage_cases()}}
}
