% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-ui.R
\name{expect_doppelganger}
\alias{expect_doppelganger}
\title{Does a figure look like its expected output?}
\usage{
expect_doppelganger(title, fig, path = NULL, ..., user_fonts = NULL)
}
\arguments{
\item{title}{The figure title is used for creating the figure file
names (all non-alphanumeric characters are converted to
\code{-}). Also, ggplot2 figures are appended with
\code{ggtitle(title)}.}

\item{fig}{A figure to test.}

\item{path}{The path where the test case should be stored, relative
to the \code{tests/figs/} folder. If \code{NULL} (the default),
the current testthat context is used to create a
subfolder. Supply an empty string \code{""} if you want the
figures to be stored in the root folder.}

\item{...}{Additional arguments passed to
\code{\link[testthat]{compare}()} to control specifics of
comparison.}

\item{user_fonts}{Named list of fonts to be aliased with font files
provided by the user rather than fonts properly installed on the
system. The aliases can be fonts from the fontquiver package,
strings containing a path to a font file, or a list containing
\code{name} and \code{file} elements with \code{name} indicating
the font alias in the SVG output and \code{file} the path to a
font file.}
}
\description{
If the test has never been validated yet, the test is skipped. If
the test has previously been validated but \code{fig} does not look
like its expected output, an error is issued. Use
\code{\link{validate_cases}()} or \code{\link{manage_cases}()} to
(re)validate the test.
}
\details{
\code{fig} can be a ggplot object, a recordedplot, a function to be
called, or more generally any object with a \code{print} method. If
a ggplot object, a dependency for ggplot2 is automatically added
(see \code{\link{add_dependency}()}).
}
\examples{
\dontrun{
disp_hist_base <- function() hist(mtcars$disp)
disp_hist_ggplot <- ggplot(mtcars, aes(disp)) + geom_histogram()

expect_doppelganger(disp_hist_base, "disp-histogram-base")
expect_doppelganger(disp_hist_ggplot, "disp-histogram-ggplot")
}
}

