% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases-ui.R
\name{manage_cases}
\alias{manage_cases}
\title{Manage visual test cases with a Shiny app}
\usage{
manage_cases(
  package = ".",
  filter = NULL,
  invert = FALSE,
  ...,
  options = list()
)
}
\arguments{
\item{package}{Package description, can be path or package
name. See \code{\link[devtools:as.package]{devtools::as.package()}} for more information.}

\item{filter}{If not \code{NULL}, only tests with file names matching this
regular expression will be executed. Matching be performed on the file
name after it has been stripped of \code{"test-"} and \code{".R"}.}

\item{invert}{should the regexp supplied to \code{filter} be inverted? Defaults to \code{FALSE}.}

\item{...}{Unused.}

\item{options}{Named options that should be passed to the \code{runApp} call
(these can be any of the following: "port", "launch.browser", "host", "quiet",
"display.mode" and "test.mode"). You can also specify \code{width} and
\code{height} parameters which provide a hint to the embedding environment
about the ideal height/width for the app.}
}
\description{
Manage visual test cases with a Shiny app
}
\seealso{
\code{\link[=vdiffrAddin]{vdiffrAddin()}}, \code{\link[=collect_cases]{collect_cases()}}, and \code{\link[=validate_cases]{validate_cases()}}
}
