% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiffr-package.R
\docType{package}
\name{vdiffr-package}
\alias{vdiffr}
\alias{vdiffr-package}
\title{vdiffr: Visual Regression Testing and Graphical Diffing}
\description{
An extension to the 'testthat' package that makes it easy to add graphical unit tests. It provides a Shiny application to manage the test cases.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://vdiffr.r-lib.org/}
  \item \url{https://github.com/r-lib/vdiffr}
  \item Report bugs at \url{https://github.com/r-lib/vdiffr/issues}
}

}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@posit.co}

Authors:
\itemize{
  \item Thomas Lin Pedersen \email{thomas.pedersen@posit.co} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})
  \item T Jake Luciani \email{jake@apache.org} (svglite)
  \item Matthieu Decorde \email{matthieu.decorde@ens-lyon.fr} (svglite)
  \item Vaudor Lise \email{lise.vaudor@ens-lyon.fr} (svglite)
}

Other contributors:
\itemize{
  \item Posit Software, PBC [copyright holder, funder]
  \item Tony Plate (svglite: Early line dashing code) [contributor]
  \item David Gohel (svglite: Line dashing code and raster code) [contributor]
  \item Yixuan Qiu (svglite: Improved styles; polypath implementation) [contributor]
  \item Håkon Malmedal (svglite: Opacity code) [contributor]
}

}
\keyword{internal}
