% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vdiffr.R
\docType{package}
\name{vdiffr-package}
\alias{vdiffr}
\alias{vdiffr-package}
\title{vdiffr: Visual Regression Testing and Graphical Diffing}
\description{
An extension to the 'testthat' package that makes it easy
    to add graphical unit tests. It provides a Shiny application to
    manage the test cases.
}
\author{
\strong{Maintainer}: Lionel Henry \email{lionel@rstudio.com}

Authors:
\itemize{
  \item Carl Sutherland (jg-imagediff library)
  \item David Hong (TwoFace library)
  \item T Jake Luciani \email{jake@apache.org} (svglite)
  \item Matthieu Decorde \email{matthieu.decorde@ens-lyon.fr} (svglite)
  \item Vaudor Lise \email{lise.vaudor@ens-lyon.fr} (svglite)
}

Other contributors:
\itemize{
  \item RStudio [copyright holder]
  \item Humble Software (jg-imagediff library) [copyright holder]
  \item  jQuery Foundation (jQuery library) [copyright holder]
  \item  jQuery contributors (jQuery library; authors listed in inst/htmlwidgets/lib/jquery-authors.txt) [contributor, copyright holder]
  \item Tony Plate (svglite: Early line dashing code) [contributor]
  \item David Gohel (svglite: Line dashing code and raster code) [contributor]
  \item Yixuan Qiu (svglite: Improved styles; polypath implementation) [contributor]
  \item Håkon Malmedal (svglite: Opacity code) [contributor]
}

}
