% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases-ui.R
\name{vdiffrAddin}
\alias{vdiffrAddin}
\title{RStudio Addin for managing visual cases}
\usage{
vdiffrAddin()
}
\description{
The package is detected by looking for the currently active
project, then for the current folder if no project is active.
}
\seealso{
\code{\link[=manage_cases]{manage_cases()}}, \code{\link[=collect_cases]{collect_cases()}}, and \code{\link[=validate_cases]{validate_cases()}}
}
